+++ dido-json/src/test/java/dido/json/JsonExamplesTest.java
@@ -10,14 +10,18 @@ import org.oddjob.arooa.convert.ArooaConversionException;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
+import java.util.Properties;
+import java.util.stream.Collectors;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -68,13 +72,9 @@ public class JsonExamplesTest {
 
         String results = lookup.lookup("results", String.class);
 
-        String expected = """
-                [
-                    { "Fruit":"Apple", "Qty":5, "Price":27.2 },
-                    { "Fruit":"Orange", "Qty":10, "Price":31.6 },
-                    { "Fruit":"Pear", "Qty":7, "Price":22.1 }
-                ]
-                """;
+        String expected = new BufferedReader(new InputStreamReader(Objects.requireNonNull(getClass()
+                .getResourceAsStream("/expected/FromToJsonArrayExample.json"))))
+                .lines().collect(Collectors.joining());
 
         JSONAssert.assertEquals(
                 results,
@@ -104,7 +104,7 @@ public class JsonExamplesTest {
         assertThat(capture.get(1).getStringNamed("Fruit"), is("null"));
 
         List<String> expectedData = Files.readAllLines(Path.of(Objects.requireNonNull(getClass()
-                .getResource("/expected/FromToJsonNullsAndNans.txt")).toURI()));
+                .getResource("/expected/FromToJsonNullsAndNansData.txt")).toURI()));
 
         assertThat(capture.stream().map(Objects::toString).toList(), is(expectedData));
 
@@ -113,6 +113,40 @@ public class JsonExamplesTest {
 
         assertThat(lookup.lookup("results.lines", List.class), is(expectedJson));
     }
+
+    @Test
+    void fromWithGsonBuilder() throws IOException, URISyntaxException {
+
+        Path workDir = Path.of("target/work");
+        Files.createDirectories(workDir);
+
+
+        Oddjob oddjob = new Oddjob();
+        oddjob.setFile(new File(Objects.requireNonNull(
+                getClass().getResource("FromToWithGsonBuilder.xml")).getFile()));
+        Properties properties = new Properties();
+        properties.setProperty("work.dir", workDir.toString());
+        oddjob.setProperties(properties);
+
+        oddjob.run();
+
+        assertThat(oddjob.lastStateEvent().getState().isComplete(), is(true));
+
+        List<String> expectedData = Files.readAllLines(Path.of(Objects.requireNonNull(getClass()
+                .getResource("/expected/FromToWithGsonBuilderData.txt")).toURI()));
+
+        List<String> actualData = Files.readAllLines(workDir.resolve("FromToWithGsonBuilderData.txt"));
+
+        assertThat(actualData, is(expectedData));
+
+        List<String> expectedJson = Files.readAllLines(Path.of(Objects.requireNonNull(getClass()
+                .getResource("/expected/FromToWithGsonBuilder.json")).toURI()));
+
+        List<String> actualJson = Files.readAllLines(workDir.resolve("FromToWithGsonBuilderOut.json"));
+
+        assertThat(actualJson, is(expectedJson));
+    }
+
 
     public static class IterableOfData implements Iterable<DidoData> {