+++ dido-json/src/test/java/dido/json/JsonStringToDataTest.java
@@ -2,6 +2,7 @@ package dido.json;
 
 import com.google.gson.GsonBuilder;
 import com.google.gson.Strictness;
+import com.google.gson.ToNumberPolicy;
 import dido.data.DidoData;
 import org.junit.jupiter.api.Test;
 
@@ -23,5 +24,19 @@ class JsonStringToDataTest {
         DidoData data = func.apply("{ \"Fruit\":\"Apple\", \"Qty\":5, \"Price\":27.2 }");
 
         assertThat(data, is(DidoData.of("Apple", 5.0, 27.2)));
+    }
+
+    @Test
+    void objectToNumberLong() {
+
+        Function<String, DidoData> func = DataInJson.with()
+                .gsonBuilder(gson -> gson.setObjectToNumberStrategy(ToNumberPolicy.LONG_OR_DOUBLE))
+                .mapFromString();
+
+        DidoData data = func.apply("{ \"Fruit\":\"Apple\", \"Qty\":5, \"Price\":27.2 }");
+
+        assertThat(data.getSchema().getTypeNamed("Qty"), is(long.class));
+
+        assertThat(data, is(DidoData.of("Apple", 5L, 27.2)));
     }
 }