+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 Dido 2.1.0
 ----------
 
-
+- Support Configuring the Gson Builder in JSON Oddjob configurations.
+- Support Long or Double parsing in JSON Oddjob configurations.
 - Fix bug where text table doesn't cope with missing indices.